Add GC sync points to various ops.
A GC sync point is a place in execution where it's safe to do a GC run.
Threads should at these points check if they are being signalled to do
a GC run and take action if so. Each allocation serves as one, but if
a thread is sat in a tight loop and never allocating, it would never
pause to participate in the GC run. Thus we mark things that cause a
jump as GC sync points.