Untangle the VM's native strings from BOOTStr.
BOOTStr is meant to be something that can box a string, with the P6str
REPR. This is distinct from the VM's own native strings, which are
factored as 6model objects but don't really count as a box. This fixes
t/qast/qast_literal.t completely.

Note that most places where BOOTStr would once have been used should
now use VMString, as shown in the many changes in this commit.