Add SC weak-reference hash.
This is how we'll look up SCs when we need to resolve them by handle.
Note this is not the common case. Also, this must not keep them alive;
it is them being referenced from compilation units that will do that.
Add removal from this hash at the point of GC free. Protected with a
mutex.