Start splitting write barrier into ref/agg forms.
The ref form is for when we are updating an address that is in the
managed body of an object, and that the GC itself can sanely track.
The agg form is for aggregates (like the VMArray and VMHash REPRs)
that may write the object reference into some memory they manage, but
that they may move as needed, meaning nothing should hold a reference
to their insides. In this case, the GC will need ask for all of the
nursery references at the point it does a mark.