Add an MVMROOT macro.
Usage is:

    MVMROOT(tc, local_variable, { ... });

For the duration of the block, the specified local variable will be
pushed as a temporary root, so even if a GC run is triggered it will
stay where it is. Boils down to the same code you'd write by hand to
call the push/pop function, but with less room for errors, less typing
and some indentation to indicate the covered region.