Avoid minor buffer overflow in say_n
t/moar/float.t fails on darwin/amd64 with an abort in __sprintf_chk.
Google tells me that this is related to a libssp data protection
library detecting a buffer overflow.

Sure enough "-127243075316.468048" is 20 characters long which puts
the null character one byte past the end of `char buf[20]`.
Allocating a single extra byte for the buffer fixes the problem.