Don't build APR's dynamic libraries
We're trying to use the static version anyway, so don't waste time
building a dynamic library.  And hopefully this will help keep some
versions of ld *cough*OS X*cough* from trying to use the dynamic
version.  (OS X is especially bad in that at runtime it searches for
where APR says its going to be installed, which is overly confusing.)