Simplify MVM_PUBLIC/MVM_PRIVATE logic.
If we go ahead with this factoring, MVM_BUILD_SHARED and MVM_SHARED
need to be passed as compiler flags.

The former will be added by Configure.pl, the latter must be
specified by library users.

Note that not supplying MVM_SHARED when using the dynamic lib
should work in principle, but comes with a performance penalty
on Win32.