Toss looking down the stack for outer.
This was a fallback, but it turns out we depended on it a lot. Now we
barely do, so it can go away. We will need a proper auto-close for a
handful of rare cases, but removing this approximation of it shows up
a couple of places we wrongly depend on such a thing, that need fixes
before putting auto-close in.