Re-order ops.
This is mostly aimed at bringing related things together, where they
ended up scattered because it's typically easier to add things at the
end. Also add a note that from now on, this order should be considered
API, with new things to be added at the end.

The order is operationally unimportant; putting related things near to
each other is simply a convenience for those working on the VM.