Fix capture objects / flattening interaction.
In the case flattening took place, we want to re-use the work, but we
must have an MVMCallsite that describes the work. Otherwise, we use an
out-of-date one, which confuses anything on receiving end of being
invoked with the capture. Future simplifications are likely possible,
but this moves things forward a step in getting Rakudo working.