Optimize root objects list in SC.
This doesn't make things better from an algorithmic point of view, but
looking up objects in SCs - and in particular, searching the SC to find
the index of an object - is a very common operation. This makes all SC
object operations a good bit faster, and easier for the compiler to do
inlining on, which is a very notable win.