MVMStaticFrame does not directly ref an MVMFrame.
Thus refs_frames can be cleared. It may have referenced them at some
point in the past, but it doesn't now. Since MVMStaticFrame was the
most costly of the refs_frames things to mark, and there were a lot
of them, this makes the intergenerational roots a bunch cheaper to
handle.