Do Just Enough so $*IN.get/$*IN.lines work.
This is *not* a good way to do, well, anything. However, pending the
upcoming review of IO stuff after this release, it hopefully gets us a
working $*IN.get and $*IN.lines. Does at least work in Windows. For
now, regress on supporting a plain \r as line ending in readline; can
put that back more easily after the future cleanup.