In P6opaque's compose(), zero the freshly allocated repr_data->unbox_slots.
Not all the slots are used, but serialize_repr_data() will serialise all of
them, and hence write out garbage to disk. This doesn't cause crashes, because
the unuused slots are never accessed, but it does conceal other problems.
valgrind reports an awful lot of warnings, concealing real problems, and the
generated bytecode files will not be byte-for-byte identical, preventing easy
consistency checking.