Add probing code to Configure.pm to learn how the compiler does 'static inline'
The compiler probing framework is intended to be generic, and is based on the
approach used by Perl 5's Configure, but with the compiler and linker flags
derived from Configure.pl's own values, which should make it portable to Win32
as well as *nix.

The probing code for static inline is a direct port of the Perl 5 code written
by Andy Dougherty, applied to Perl 5 as commit 17a6c8e38505fd8d in July 2010.