Fix a source of frame leaks in continuations.
When we GC a continuation that sliced off some active handlers, and
those handlers still reference frames, make sure we decrement the
frame reference count. Can happen for a gather/take that does not
ever complete its execution. So, not the big leak affecting all of
the gather/takes, but a needed fix anyway.