Fix "on stack" continuation ref-count handling.
When a frame is on the current call stack, it gets its reference count
incremented by 1 to indicate it's active at the very least as a result
of currently being running. When we take a continuation, to make sure
frame lifetimes are handled correctly, we just decrement this count as
we slice frames away. Similarly, continuation invoke should re-instate
the count as we splice the frames back onto the active call stack.
Fixes gather/take memory leak reported by TimToady++.