Add setmultispec mechanism.
When we have onlystar protos with simple signatures, there's no reason
to call the proto, to have it wrap the args in an object to then use
to do a lookup in a multi cache. Provide a mechanism to try the cache
right away, falling back to the proto (which may populate the cache).