Change the way lexotic is handled.
Make it not imply a frame reference, at the cost of a little more work
at return time. Then also avoid allocating one per frame that needs it.
This leads to a 10% saving on "for 1..1000000 { $i++ }" in Rakudo; we
can likely also add a spesh op to make obtaining a lexotic cheaper
still.