try to optimize islist/isnum/... (not helpful yet)
the problem is that islist and friends also include a null
check on their argument. At best, we could - if we only know
the type - turn these ops into a negated "isnull" op, but that
doesn't exist. So we have to wait for spesh's ability to
allocate new registers and add new operations on those.