Add lexical lookup lifetime hinting ops.
Allow us to convey that a given lexical lookup can be relied on to
always resolve to the same thing, or can be relied on to resolve to
the same thing per invocant type. The former is useful for setting
sub lookups, the latter for type variables. For now don't do anything
extra, but spesh will be able to improve on this.