Support for ARMv6. Don't try to read unaligned doubles or long longs.
Specifically, ARMv6 will SIGBUS for an unaligned double, and generate a CPU
exception for an unaligned long long, which Linux defaults to catching with a
fixup handler. This means that we can't probe for unaligned long long reads,
as ARMv6 Linux systems would pass the probe, but we wouldn't realise we had
chosen a horribly inefficient approach.

This commit replaces the macros GET_I64 and GET_N64 with inline functions
MVM_BC_get_I64 and MVM_BC_get_N64, and removes the unused macro GET_U64.
It adds probing code to determine CPU architecture, which is hopefully robust
against all the crazy that OSes try to throw at us. The probing code is ready
for platforms where unaligned 32 bit reads are bad, but the C code is not yet,
as we don't yet pass tests on any that I have access to.