Iterate to fixed point when finding unused instrs.
Before we just did a backwards propagation as we went back up the
dominator tree. But that didn't catch enough. Switch to just keeping
iterating doing removals until we're done. May be more optimal ways
to walk BBs in order to speed up convergence. Either way, the results
with doing the analysis this way are better.