Add Makefile rules to generate pre-processed source.
Someone awkwardly, @objflags@ can contain -DMVM_BUILD_SHARED, which affects the
pre-processed source, so for correctness that needs to be in the CPP command
line, even if it means that we're also sucking in linker-only flags.

It seems that .i is used everywhere, so KISS implies no need for it to be a
setup option.