Add Makefile rules to generate assembler.
Again, as @objflags@ can contain -DMVM_BUILD_SHARED it needs to be in the
command line.

Win32 doesn't quite work out, since the /GL used to enable link time
code generation cause the compiler to decide it won't bother spitting
out the .asm file. Hack that flag out of the Makefile and an assembly
file is, indeed, produced.