Pre-invalidate all deopt target addresses.
This prevents us accidentally having them match up. This can happen if
we set them when producing the logging bytecode, but then toss chunks
of code (or just instructions) in the optimization phase and so never
encounter the deopt annotation again when doing code generation of the
optimized code. This could lead to bogus entries being used to deopt,
essentially placing the interpreter at an arbitrary instruction.