On big endian platforms, swap endiannes in ensure_bytes() while validating.
Astoundingly, this is all that is needed to get NQP to build on 64 bit POWER,
and pass all tests other than nativecall. (Which the backtrace suggests is
because dyncall is using ppc32). Oh, and Rakudo builds and passes all tests.

Certainly it's not the most efficient thing, compiling endian-swapped bytecode
and serialisation data in memory, only to swap it back immediately to run it.
"The marvel is not that the bear dances well, but that the bear dances at all."