Change the way we store SCs in object headers.
We used to store a pointer directly to the SC. Now we store a pair
of 32-bit integers: an index into a new array where we can look up
the SC, and the index the object lives at in the SC, if known. This
second index is not being set up everywhere consistently yet; this
patch does what's needed to switch over to the new header layout
and be able to build NQP/Rakudo.