Add struct MVMSerializationIndex to hold large STable indexes.
Continue to store indexes that fit directly in the union, but for larger
values, allocate a struct MVMSerializationIndex, store the indexes there,
and use the union to store a pointer to it.

For now, we don't have overflow checks on values for the global SC index, as
the highest seen value is 149, which is considerably smaller than 65536.

Currently the code leaks all memory allocated. That will be fixed in the next
commit. The handling of STable collection is also LTA, to put it politely.

For now, force 32 and 64 bit platforms to exercise the oversize code paths.
To torture them properly, set the oversize threshold to be 255, not 65535,
and add an extra debugging field to try to catch problems. This code is
intentionally left in for now, to make it easier to debug problems.