make exception handlers aware of labels
Frame handlers have now a register which can hold a label in case we have a
labeled loop. When an exception is thrown and the category is MVM_EX_CAT_LABELED,
we check for object identity of the thrown payload and the label in the register.
That means we will have handlers for e.g. 'redo', and this handler will care
about 'redo' exceptions without a label, and also will care if the label fits.