Don't return to upper frame when deopting from JIT
This fixes a puzzling issue in which deopt was called but
the scripts failed anyway. The cause was that - up until now - the
special 'jit enter' opcode assumed that return from the jit code
is equal to return from the frame, which doesn't hold in deopt
(because deopt restores the frame for you).

This assumption also wouldn't have held once we had support for
invocations, so good riddance for it :-)