Add option to configure lib target directory
The build system does not permit to configure the lib target directoriy,
this is required for Debian packaging with multi-arch support[1].

The new “Configure.pl” “libdir” option control the install path for the
library.  The default is “@prefix@/lib/” on POSIX toolchain and
“@bindir@” on MSVC.

The new “libdir” replace both the “shareddir” and “staticdir” options.

Footnotes:
[1]  https://www.debian.org/doc/manuals/maint-guide/advanced.en.html#multiarch