Refactor JIT graph building
I refactored the JIT graph building to use a 'JitGraphBuilder'
structure. In this way, the graph builder can 'consume'
more than one spesh instruction at a time, which is very useful
for many reasons. One of these reasons is that it allows us to
build a single 'master invoke node'.