Cope with cross-deopt-point usage analysis.
If a register gets used after a deopt point, give its usage an extra
bump. This means we won't kill a write to it when doing dead code
elimination. Even if we don't actually read it later in the optimized
code, we still need things set up right for if we deopt. This extra
care will allow us to be a little more aggresive in decrementing usage
in various other situations.