Significantly overhaul the strings code.
Introduce a grapheme iterator, along with a codepoint iterator (that
is just an alias to the same thing). Go with a flat approach to the
strands, which greatly simplifies traversal of a string and should do
a little better CPU-cache wise. Actually use strands in concatenation,
substring, joining, and repeat operations.

This patch does enough to get NQP and Rakudo building and passing the
sanity tests with the new approach; spectest looks good too. Still a
good amount of optimization and improvements possible, which will be
easier now the code is cleaner.