Fix issue with zero-arity named multis
Named arguments do not count for the arity in the
MVMMultiCache, however they do not fit in the
'zero arity' slot either. Consequently, when supplied
with a callsite without positionals but with named arguments,
the cache algorithm searched for matches in a negative offset
of the arity entries array. This fixes that by refusing to
deal with such callsites.