Cope better with a frame being re-OSR'd.
If this happened, the frame may have been resized already. The upshot
of this is that we didn't do the "allocate zeroed and copy" sequence
that ensures the environment is empty for the inlinees. Thus, old and
bogus pointers could end up being grabbed from the previous OSRing.
Fixing this clones #116.