cur_callsite is no longer top of stack
MVM_frame_find_invokee_multi_ok expects to take the
address of the current callsite. I used to push the
pointer to this object on the stack, so that the stack
pointer was in fact the address of this object. However,
I don't use push and pop anymore, instead using a statically
allocated stack frame of 96 bytes large. Since this
means the cur_callsite pointer is stored somewhere else, the
3rd argument is gibberish. Fix by loading the current location
into the 3rd argument.

nwc10++ and timotimo++ for getting good backtraces.