Don't compile frames that have handlers
Handlers and the JIT don't play so nicely together, yet.
This is because searching a frame handler looks for the
current bytecode offset, which is all off in the JIT
because it uses a magic bytecode.

So I'll disable the handlers for now. This fixes a long-standing
issue whereby adding the iter ops would case infinite loops and
other weird behavior.