Split out part of frame deserialization.
The split out part will be able to happen lazily, the first time we
need it. (At present that won't be much of a win as we touch many of
the frames at startup to install static lexical information; the plan
is to move this information into the bytecode file also).