Redo label handling for deopt and handler support
Instead of labels and offset, I now keep just indexes into
the label array and indexes into the deopt array. Believe it
or not, this makes quite a few things simpler, such as removing
the necessity of a different OSR label array and a deopt_all label
array. The goal is to make the construction of handler labels
possible during graph construction, as well as inline labels.