Generalise throwish control.
Sometimes, the throwish op may throw to the same frame the source.
The simplest way of dealing with that by far is letting the
interpreter start the JIT code in the right place. But in that case
the invokish control guard will not do the right thing, so it has
to be replaced by two different guards.