Fix newlexotic
Newlexotic normally takes the bytecode offset of the frame
handler goto label, but this is transformed into a pointer to
the basic block during spesh graph creation. Instead of looking
up the correct value using some hack, pass the label instead.

Because the goto annotation is attached to an instruction and
the instruction points to a block, the labels of these would diverge
if the annotated instruction was not the first of the basic block,
as would happen if there where PHI nodes before. So I've changed the
labeling code to disregard PHI nodes when labeling an instruction.