Enable code objects to be lazily deserialized.
We now can attach the indexes to a MAST::Frame, and keep them around
until such a time as we need the code object. The goal is that, in
conjunction with some other improvements, we might be able to avoid
deserializing many code objects in Perl 6, along with signatures,
parameters, etc.