Fix bug in handlers
When a frame is JITted, the exception handling logic must
search for JIT handlers rather than regular handlers because
the interpreter bytecode offset. Except that this is again
meaningless if that code has never ran, such as when the frame
was compiled but has not been entered yet. So we must check
for that, as well.

I've also added code to automatically load the right label
to the interpreter whenever we cross a frame handler barrier,
so that adhoc exceptions should Just Work.