[WIP] implement multi-char input line separators
This implemements line seps on a grapheme level that are not just one character long.
These separators are used when cutting lines from an already filled buffer, but they
are also used to fill a buffer by requesting more data from a stream until we find a
separator.
Parts of this code already handle multiple separators, others do not. Also, when we
have more than one separator we usually want to prefer the longest matching separator.
To achieve that we keep a list of possible remainders that need get chopped off the
next line, in case we have separators that start with the same chars. Like \n would be
a remainder for the separators [ \r\n | \r ], when we only hit \r yet.
We also need an op that takes a list_s, since nqp::setinputlinesep() only takes one
string.
Note that this patch still leaks as hell, e.g. the list of input separators and
remainders. Though it spectests fine, which is always nice.