implement nqp::closefhi which retuns the exit code of a pipe
nqp::openpipe returns a file handle one can read from, but since such an
opened pipe is actually a spawned process we also need its exit code to
know if it run cleanly. Use nqp::closefhi on such a file handle if you
are interested in the exit code, nqp::closefh if you are not.