attempt to fix a non-alloc+free #154
In case we have no input praphs in a string join, we will return an empty string constant but
we allocate a result string first, where its body.storage.any pointer could be garbage.
Later we gc_free that result string that was never used, and trip over that garbage pointer.
Now we set this pointer to NULL before returning from that function.