Avoid a pointer getting outdated on the stack.
If the box operations allocates, then a now-outdated `key` pointer was
already copied onto the stack to pass to the bind_key. Thus it came to
work on this now-moved object body. This may have been the cause of the
various mostly-on-OSX failures, given it was discovered using the OSX
guardmalloc while investigating the issues there.