trap epsilons to 0
The NFA construction code could accidentally introduce a state with no
edges, which under epsilon removal produced an epsilon that forwarded to
state 0, which doesn't work so well.  This bug is now fixed, but until
we rebootstrap, we rely on moar's NFA to ignore null epsilons for later
stages optimizing earlier stages' NFAs.