Big cleanup of signed/unsigned mess.
Turns out that using anything but char * for things that are C level
and string-ish goes against enough C culture to be a real pain. So
standardize on that. Clears up almost all the remaining warnings in
clang, and the overall number of casts has probably decreased some
along the way.