Start getting NativeRef REPR fleshed out.
We put the kind of referenced thing and the primitive type as REPR
data. This means there'll be a minor explosion of types for native
references, but they align well with the way we'll want to do type
specialization, so it's worth it. They'll also be invisible to most
users (except those who go digging).