Simplify facts known value handling.
If we know a value is an integer constant as a fact, it's really not
important for optimization precisely what size register or constant
operand we obtained it from. This patch tosses the sized variants and
fixes the tiny number of places that used to store to them. This will
hopefully avoid introducing BE bugs in the future.