Re-enable lazy deserialization.
Provide a #define for easily turning it on/off also. We'll see what
the ecosystem fall-out is; even if there is any, we really want to
have this on for the startup and memory use, so we should discover
and debug the issues ahead of the next release. The pre-comp tests
in the spectest suite are no worse with this turned on.