Initial, simple use of the frame gen 2 only flag.
This just uses it to elide scan_registers, which is the easiest thing
we can do with it though with the least benefit. It will, however, let
us see if there are any bugs with the frame write barrier stuff so far,
and saves enough time to pay the cost of the barrier.

The next step will be to raise it to cover the MVMFrame overall; the
end goal is to eliminate refs_frames on REPRs, and vastly reduce the
size of the inter-generational set.