First step for parameterized type serialization.
This gets us serializing the relation with the parametric type as well
as the parameters used. On deserialization, this relationship is then
put back in place. What's missing is parameterization interning, but
code to detect what things need an entry in the intern section is in
place; what remains is to write such a table and process it at the
point of deserialization.