Inline assert_can_read(), which is called 7,133,540 times for perl6-m -e 1
assert_can_read() is a very small function, but 15th highest by Ir count in
callgrind's output for perl6-m -e 1. Hence the overhead of making the function
call is likely to be similar in size to the function's body.

cachegrind reports that inlining it reduces I refs from 515,809,288 to
509,668,666, ie a 1% reduction in CPU instructions needed to run perl6-m -e 1