Verify that the packed object SC id/index pair match the data section values.
(Testing this relies on the fact that currently no part of setting compilation
happens to have values so big that they overflow the packed format. Once the
testing code is removed, that assumption is removed, and we will still be able
to store stupid-big values.)

Also, write garbage into the values in the object table, to verify that
nothing is reading them.