Verify that we can write a packed object SC id/index pair that matches.
We also verify that we've correctly identified all the locations of reader
code that needs to cope with old and new formats, because the new format is
now a different size from the old format (which stage0 uses).

Again, testing this relies on the fact that currently no part of setting
compilation happens to have values so big that they overflow the packed
format.