MVM_serialization_write_str() writes string offsets as 15 or 31 bit values.
So far, all string offsets that we generate are much less than 2**15, so
we're writing out a lot of 0 bytes if we store then as a fixed 32 bit value.
However, switching to the varint format is not viable, as it significantly
increases the CPU work needed to load the setting. So instead, use a 2 or 4
byte format, which for the setting will always be exactly one 16 bit read.

For now, force the code to always use the longer form, to verify that the
reader works.