Note that deserialize_method_cache_lazy() is cruel and unforgiving.
As the comment said, deserialize_method_cache_lazy() stashes what we need to
eserialize the method cache lazily later, and then skips over it.

As the comment now adds:

This function is cruel and unforgiving if you change other parts of the
serialization format, but don't remember (or realise) that you need to update
its idea of sizes. Its "failure" mode is silent, and everything still passes
tests. Only if you benchmark do you realise that everything takes longer,
because the lazy paths are now no longer taken.