Implement re-use of synthetic graphemes.
This is needed to get string equality correct. We use a partially
lock-free trie to achieve this. Having read the base of the trie,
any thread is free to traverse it without having to acquire a lock
(that is, reads are always lock free). Only additions need the lock,
and it exists to serialize the additions. We always copy nodes that
are modified, and never modify things in place; we schedule the memory
to be freed at the next global safe point (at which point we know that
no thread could possibly be reading any version of the trie).