Fix SSA crash when huge number of basic blocks.
When we have a lot of basic blocks, we might end up needing to produce
huge PHI nodes. This was exposed in real world code of only 8KB when a
huge alternation was written in a regex. This fix both bumps the size
of the field holding operand count so we can handle such large graphs,
and adds a check so we panic rather than SEGV in the (rather unlikely)
event we ever exceed the new limit.