Inititialise `operand_size` in MVM_bytecode_dump()
The variable `operand_size` could be used without first being set (namely in
the assignment on line 324), thus it is being initialised to avoid
uninitialised accesses.  The value chosen here is 0, which seems an
appropriate choice given the context (there isn't really an op here, so the
current op doesn't need to be incremented).  Another initial value could be
more appropriate, hence this change should be reviewed before being merged.