Optimize temporary root handling.
Only do sanity checks when we switch them on; they've caught very few
bugs so far, so it doesn't pay to have them on all the time. This in
turn means they are well within the size where they make sense to do
inlining; mark with MVM_STATIC_INLINE to be sure of it.