Dual-purpose comp unit string heap as SC one.
This means we don't have to construct an array of strings in the order
the deserialization blob wants them, and can instead just point the
deserialization at the string heap of the compilation unit. When the
NQP change to take advantage of this is applied, we knock 60KB off the
size of CORE.setting, and save 2% off startup time.