Thread ID handling cleanup.
Differentiate VM-level ID from native thread ID. Make nqp::threadid
be the former, as on JVM, so we can even know it before the thread is
started. Make sure we give a thread ID early enough to answer the
question. Also, never block on a request for a thread ID. We don't yet
have an API to get the native thread ID, but keep the code around as
we'll likely get one some day.